requirements.txt
mediapipe==0.10.21
opencv-python
numpy
flask-cors
